Rename deposits to current spend
