+++ src/Dashboard.js
@@ -19,7 +19,7 @@ import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import NotificationsIcon from '@material-ui/icons/Notifications';
 import { mainListItems, secondaryListItems } from './listItems';
 import Chart from './Chart';
-import Deposits from './Deposits';
+import Deposits from './CurrentSpend';
 import Orders from './BudgetHistory';
 
 function Copyright() {
